+++ src/main/kotlin/com/markskelton/ThemeConstructor.kt
@@ -15,6 +15,7 @@ import com.markskelton.settings.Groups.*
 import com.markskelton.settings.ThemeSettings
 import com.markskelton.settings.toGroup
 import com.markskelton.settings.toGroupStyle
+import com.markskelton.utils.readXmlInputStream
 import groovy.util.Node
 import groovy.util.XmlNodePrinter
 import groovy.util.XmlParser
@@ -215,21 +216,10 @@ object ThemeConstructor {
   }
 
   private fun getEditorXMLTemplate(): Node =
-    this::class.java.getResourceAsStream("/templates/one-dark.template.xml").use { input ->
-      val inputSource = InputSource(InputStreamReader(input, "UTF-8"))
-      val parser = XmlParser(false, true, true)
-      parser.setFeature("http://apache.org/xml/features/nonvalidating/load-external-dtd", false)
-      parser.errorHandler = object : ErrorHandler {
-        override fun warning(exception: SAXParseException?) {}
-
-        override fun error(exception: SAXParseException?) {}
-
-        override fun fatalError(exception: SAXParseException) {
-          throw exception
-        }
+    this::class.java.getResourceAsStream("/templates/one-dark.template.xml")
+      .use { input ->
+        readXmlInputStream(input)
       }
-      parser.parse(inputSource)
-    }
 
   private fun getColorPalette(themeSettings: ThemeSettings): ColorPalette {
     val selectedPalette = if (themeSettings.isVivid) "vivid" else "normal"
