+++ src/main/kotlin/com/markskelton/settings/ThemeConfigListener.kt
@@ -4,7 +4,7 @@ import com.intellij.util.messages.Topic
 import java.util.*
 
 val THEME_CONFIG_TOPIC: Topic<ThemeConfigListener> =
-  Topic(ThemeConfigListener::class.java)
+  Topic("One Dark Settings", ThemeConfigListener::class.java)
 
 interface ThemeConfigListener : EventListener {
   fun themeConfigUpdated(themeSettings: ThemeSettings)
